Move annealing functions to optim.py
